Remove dead e2e code, combine home pages
